+++ src/prime_rl/trainer/batch.py
@@ -6,7 +6,6 @@ from jaxtyping import Bool, Float, Int
 from torch import Tensor
 from transformers.tokenization_utils import PreTrainedTokenizer
 
-from prime_rl.trainer.rl.data import MicroBatch
 from prime_rl.utils.vf import Rollout
 
 
@@ -16,6 +15,19 @@ class BatchSample(TypedDict):
     loss_mask: Bool[Tensor, "seq"]
     advantages: Float[Tensor, "seq"]
     inference_logprobs: Float[Tensor, "seq"]
+
+
+class MicroBatch(TypedDict):
+    # Token level
+    input_ids: Int[Tensor, "batch seq"]
+    position_ids: Int[Tensor, "batch seq"]
+    advantages: Float[Tensor, "batch seq"]
+    inference_logprobs: Float[Tensor, "batch seq"]
+    loss_mask: Bool[Tensor, "batch seq"]
+
+    # Batch level
+    temperature: float
+    lora_cu_offsets: Int[Tensor, "n_loras"]
 
 
 def prepare_sample(
